Move modal map open functionality to hook
